Crate rusoto_mobile

AWS Mobile Service provides mobile app and website developers with capabilities required to configure AWS resources and bootstrap their developer desktop projects with the necessary SDKs, constants, tools and samples to make use of those resources.

If you're using the service, you're probably looking for MobileClient and Mobile.

Structs

BundleDetails
The details of the bundle.

CreateProjectRequest
Request structure used to request a project be created.

CreateProjectResult
Result structure used in response to a request to create a project.

DeleteProjectRequest
Request structure used to request a project be deleted.

DeleteProjectResult
Result structure used in response to request to delete a project.

DescribeBundleRequest
Request structure to request the details of a specific bundle.

DescribeBundleResult
Result structure contains the details of the bundle.

DescribeProjectRequest
Request structure used to request details about a project.

DescribeProjectResult
Result structure used for requests of project details.

ExportBundleRequest
Request structure used to request generation of custom SDK and tool packages required to integrate mobile web or app clients with backed AWS resources.

ExportBundleResult
Result structure which contains link to download custom-generated SDK and tool packages used to integrate mobile web or app clients with backed AWS resources.

ExportProjectRequest
Request structure used in requests to export project configuration details.

ExportProjectResult
Result structure used for requests to export project configuration details.

ListBundlesRequest
Request structure to request all available bundles.

ListBundlesResult
Result structure contains a list of all available bundles with details.

ListProjectsRequest
Request structure used to request projects list in AWS Mobile Hub.

ListProjectsResult
Result structure used for requests to list projects in AWS Mobile Hub.

MobileClient
A client for the AWS Mobile API.

ProjectDetails
Detailed information about an AWS Mobile Hub project.

ProjectSummary
Summary information about an AWS Mobile Hub project.

Resource
Information about an instance of an AWS resource associated with a project.

UpdateProjectRequest
Request structure used for requests to update project configuration.

UpdateProjectResult
Result structure used for requests to updated project configuration.

Enums

CreateProjectError
Errors returned by CreateProject

DeleteProjectError
Errors returned by DeleteProject

DescribeBundleError
Errors returned by DescribeBundle

DescribeProjectError
Errors returned by DescribeProject

ExportBundleError
Errors returned by ExportBundle

ExportProjectError
Errors returned by ExportProject

ListBundlesError
Errors returned by ListBundles

ListProjectsError
Errors returned by ListProjects

UpdateProjectError
Errors returned by UpdateProject

Traits

Mobile
Trait representing the capabilities of the AWS Mobile API. AWS Mobile clients implement this trait.