Trait rusoto_mobile::Mobile

pub trait Mobile {
    fn create_project(
        &self, 
        input: CreateProjectRequest
    ) -> RusotoFuture<CreateProjectResult, CreateProjectError>;

    fn delete_project(
        &self, 
        input: DeleteProjectRequest
    ) -> RusotoFuture<DeleteProjectResult, DeleteProjectError>;

    fn describe_bundle(
        &self, 
        input: DescribeBundleRequest
    ) -> RusotoFuture<DescribeBundleResult, DescribeBundleError>;

    fn describe_project(
        &self, 
        input: DescribeProjectRequest
    ) -> RusotoFuture<DescribeProjectResult, DescribeProjectError>;

    fn export_bundle(
        &self, 
        input: ExportBundleRequest
    ) -> RusotoFuture<ExportBundleResult, ExportBundleError>;

    fn export_project(
        &self, 
        input: ExportProjectRequest
    ) -> RusotoFuture<ExportProjectResult, ExportProjectError>;

    fn list_bundles(
        &self, 
        input: ListBundlesRequest
    ) -> RusotoFuture<ListBundlesResult, ListBundlesError>;

    fn list_projects(
        &self, 
        input: ListProjectsRequest
    ) -> RusotoFuture<ListProjectsResult, ListProjectsError>;

    fn update_project(
        &self, 
        input: UpdateProjectRequest
    ) -> RusotoFuture<UpdateProjectResult, UpdateProjectError>;
}

Trait representing the capabilities of the AWS Mobile API. AWS Mobile clients implement this trait.

Required methods

fn create_project(
&self,
input: CreateProjectRequest
) -> RusotoFuture<CreateProjectResult, CreateProjectError>

Creates an AWS Mobile Hub project.

fn delete_project(
&self,
input: DeleteProjectRequest
) -> RusotoFuture<DeleteProjectResult, DeleteProjectError>

Delets a project in AWS Mobile Hub.

fn describe_bundle(
&self,
input: DescribeBundleRequest
) -> RusotoFuture<DescribeBundleResult, DescribeBundleError>

Get the bundle details for the requested bundle id.

fn describe_project(
&self,
input: DescribeProjectRequest
) -> RusotoFuture<DescribeProjectResult, DescribeProjectError>

Gets details about a project in AWS Mobile Hub.

fn export_bundle(
&self,
input: ExportBundleRequest
) -> RusotoFuture<ExportBundleResult, ExportBundleError>

Generates customized software development kit (SDK) and or tool packages used to integrate mobile web or mobile app clients with backend AWS resources.

fn export_project(
&self,
input: ExportProjectRequest
) -> RusotoFuture<ExportProjectResult, ExportProjectError>

Exports project configuration to a snapshot which can be downloaded and shared. Note that mobile app push credentials are encrypted in exported projects, so they can only be shared successfully within the same AWS account.

fn list_bundles(
&self,
input: ListBundlesRequest
) -> RusotoFuture<ListBundlesResult, ListBundlesError>

List all available bundles.

fn list_projects(
&self,
input: ListProjectsRequest
) -> RusotoFuture<ListProjectsResult, ListProjectsError>

Lists projects in AWS Mobile Hub.

fn update_project(
&self,
input: UpdateProjectRequest
) -> RusotoFuture<UpdateProjectResult, UpdateProjectError>

Update an existing project.

Loading content...